# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Inventario",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
        "/api/categories": {
            "get": {
                "summary": "Obtener todos las categorías",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "200": {
                        "description": "Lista de categorias"
                    },
                    "500": {
                        "description": "Categorias no encontrados"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Crear una nueva categoría",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Category Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion Nueva categoría"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Obtener una categoría por ID",
                "description": "Obtiene una categoría específica por su ID.",
                "operationId": "c68e76d323c008827a9e47398b1583de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Category Name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Descripcion Nueva categoría"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Category not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Actualizar una categoría existente",
                "operationId": "0e686b2748211cc688333ed705dc111f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nueva categoría"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion Nueva categoría"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada exitosamente"
                    },
                    "400": {
                        "description": "Error al actualizar la categoría"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Eliminar una categoría existente",
                "operationId": "4c12e22a7f8c617bd83bfa1fdc05428d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Categoría eliminada exitosamente"
                    },
                    "400": {
                        "description": "Error al eliminar la categoría"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "summary": "Obtener todos los productos",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "Lista de productos"
                    },
                    "500": {
                        "description": "Productos no encontrados"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "quantity",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Product Name"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion Nuevo producto"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "quantity": {
                                        "type": "decimal",
                                        "example": "10"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "producto creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener un producto por ID",
                "description": "Obtiene un producto específica por su ID.",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "producto obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Product Name"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Product Description"
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "quantity": {
                                            "type": "decimal",
                                            "example": "10"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Product not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar un producto existente",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "quantity",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Update producto"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripcion update producto"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "quantity": {
                                        "type": "decimal",
                                        "example": "10"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "producto actualizada exitosamente"
                    },
                    "400": {
                        "description": "Error al actualizar la producto"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar un producto existente",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "producto eliminado exitosamente"
                    },
                    "400": {
                        "description": "Error al eliminar el producto"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Iniciar sesión de usuario",
                "operationId": "0fc9610184083ac6dddab2c63139225a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Correo electrónico del usuario",
                                        "type": "string",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "description": "Contraseña del usuario",
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJodHRwczovL2xvY2FsaG9zdDo4MDAwL2FwaS9sb2dpbiIsImlhdCI6MTY0NzIyMjExNSwiZXhwIjoxNjQ3MjI1NzE1LCJuYmYiOjE2NDcyMjIxMTUsImp0aSI6IjA1ZWVmYmIwLTQ5NDItNGE1YS05ZGJkLTNlOWQ2MzVhYTY2MCJ9.WLgQjAqEn0KMR6ytVkRIfO4mcybS2lWDK8IQ-dJK4O8"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}